Basic SNN implementation
This is an implementation of the basic SNN architecture presented in the snnTorch tutorials. It's not obvious from the code in this notebook because this represents the finalized, most efficient version that uses all the snnTorch packages/shortcuts, but the modifications to time constants will probably need to happen in the snn.Leaky unit based on what I've read in the earlier tutorials deriving a single LIF neuron model. I am leaving this architecture in its most basic state for now and will implement modifications in the Optimizing-time-constants notebook just in case the network breaks.